Add base metrics span processor class
